+++ test/directions_test.dart
@@ -134,6 +134,17 @@ Future<void> main() async {
                 'https://maps.googleapis.com/maps/api/directions/json?origin=Toronto&destination=Montreal&departure_time=$d&key=$apiKey'));
       });
 
+      test('departure_time with now', () {
+        expect(
+            directions.buildUrl(
+                origin: 'Toronto',
+                destination: 'Montreal',
+                departureTime: 'now'),
+            equals(
+              'https://maps.googleapis.com/maps/api/directions/json?origin=Toronto&destination=Montreal&departure_time=now&key=$apiKey',
+            ));
+      });
+
       test('arrival_time', () {
         int d = 1343641500;
         expect(
@@ -250,6 +261,21 @@ Future<void> main() async {
                 ]),
             equals(
                 'https://maps.googleapis.com/maps/api/directions/json?origin=Toronto&destination=Montreal&waypoints=optimize:true|Paris|42.2,21.3|place_id:ChIJ3S-JXmauEmsRUcIaWtf4MzE|enc:gfo}EtohhU:&key=$apiKey'));
+      });
+
+      test('alternatives', () {
+        expect(
+            directions.buildUrl(
+                origin: 'Toronto', destination: 'Montreal', alternatives: true),
+            equals(
+                'https://maps.googleapis.com/maps/api/directions/json?origin=Toronto&destination=Montreal&alternatives=true&key=$apiKey'));
+        expect(
+            directions.buildUrl(
+                origin: 'Toronto',
+                destination: 'Montreal',
+                alternatives: false),
+            equals(
+                'https://maps.googleapis.com/maps/api/directions/json?origin=Toronto&destination=Montreal&alternatives=false&key=$apiKey'));
       });
     });